Move Product's embeded domain classes to Product.groovy
